+++ Frontend/src/app/layout.tsx
@@ -22,7 +22,7 @@ export default function RootLayout({
   return (
     <html lang="en" className={roboto.className}>
       <body className="flex relative h-screen overflow-hidden">
-        <SessionProvider>
+        <SessionProvider refetchInterval={0} refetchOnWindowFocus={false}>
           <SideBarWrapper />
           <main className="flex-1 h-screen overflow-hidden">{children}</main>
         </SessionProvider>